+++ src/electron/capture/CaptureHelper.ts
