+++ src/electron/capture/CaptureOptions.ts
